refactor testcase
