+++ src/test/java/com/alibaba/json/bvt/serializer/filters/NameFilterTest_byte.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt.serializer;
+package com.alibaba.json.bvt.serializer.filters;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -10,12 +10,16 @@ import com.alibaba.fastjson.serializer.JSONSerializer;
 import com.alibaba.fastjson.serializer.NameFilter;
 import com.alibaba.fastjson.serializer.SerializeWriter;
 
-public class NameFilterTest_double extends TestCase {
+public class NameFilterTest_byte extends TestCase {
 
     public void test_namefilter() throws Exception {
         NameFilter filter = new NameFilter() {
 
             public String process(Object source, String name, Object value) {
+                if (value != null) {
+                    Assert.assertTrue(value instanceof Byte);
+                }
+                
                 if (name.equals("id")) {
                     return "ID";
                 }
@@ -54,7 +58,7 @@ public class NameFilterTest_double extends TestCase {
         serializer.getNameFilters().add(filter);
 
         Map<String, Object> map = new HashMap<String, Object>();
-        map.put("id", 0);
+        map.put("id", (byte) 0);
         serializer.write(map);
 
         String text = out.toString();
@@ -63,14 +67,14 @@ public class NameFilterTest_double extends TestCase {
 
     public static class Bean {
 
-        private double  id;
+        private byte    id;
         private String name;
 
-        public double getId() {
+        public byte getId() {
             return id;
         }
 
-        public void setId(double id) {
+        public void setId(byte id) {
             this.id = id;
         }
