+++ src/test/java/com/alibaba/json/bvt/serializer/filters/NameFilterTest_char.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt.serializer;
+package com.alibaba.json.bvt.serializer.filters;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -10,17 +10,15 @@ import com.alibaba.fastjson.serializer.JSONSerializer;
 import com.alibaba.fastjson.serializer.NameFilter;
 import com.alibaba.fastjson.serializer.SerializeWriter;
 
-public class NameFilterTest_boolean extends TestCase {
+public class NameFilterTest_char extends TestCase {
 
     public void test_namefilter() throws Exception {
         NameFilter filter = new NameFilter() {
 
             public String process(Object source, String name, Object value) {
-                if (value != null) {
-                    Assert.assertTrue(value instanceof Boolean);
-                }
-                
+
                 if (name.equals("id")) {
+                    Assert.assertTrue(value instanceof Character);
                     return "ID";
                 }
 
@@ -37,7 +35,7 @@ public class NameFilterTest_boolean extends TestCase {
         serializer.write(a);
 
         String text = out.toString();
-        Assert.assertEquals("{\"ID\":false}", text);
+        Assert.assertEquals("{\"ID\":\"0\"}", text);
     }
 
     public void test_namefilter_1() throws Exception {
@@ -58,23 +56,23 @@ public class NameFilterTest_boolean extends TestCase {
         serializer.getNameFilters().add(filter);
 
         Map<String, Object> map = new HashMap<String, Object>();
-        map.put("id", true);
+        map.put("id", '0');
         serializer.write(map);
 
         String text = out.toString();
-        Assert.assertEquals("{\"ID\":true}", text);
+        Assert.assertEquals("{\"ID\":\"0\"}", text);
     }
 
     public static class Bean {
 
-        private boolean id;
-        private String  name;
+        private char   id = '0';
+        private String name;
 
-        public boolean isId() {
+        public char getId() {
             return id;
         }
 
-        public void setId(boolean id) {
+        public void setId(char id) {
             this.id = id;
         }
