+++ src/test/java/com/alibaba/json/bvt/serializer/filters/NameFilterTest_double.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt.serializer;
+package com.alibaba.json.bvt.serializer.filters;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -10,15 +10,13 @@ import com.alibaba.fastjson.serializer.JSONSerializer;
 import com.alibaba.fastjson.serializer.NameFilter;
 import com.alibaba.fastjson.serializer.SerializeWriter;
 
-public class NameFilterTest_char extends TestCase {
+public class NameFilterTest_double extends TestCase {
 
     public void test_namefilter() throws Exception {
         NameFilter filter = new NameFilter() {
 
             public String process(Object source, String name, Object value) {
-
                 if (name.equals("id")) {
-                    Assert.assertTrue(value instanceof Character);
                     return "ID";
                 }
 
@@ -35,7 +33,7 @@ public class NameFilterTest_char extends TestCase {
         serializer.write(a);
 
         String text = out.toString();
-        Assert.assertEquals("{\"ID\":\"0\"}", text);
+        Assert.assertEquals("{\"ID\":0}", text);
     }
 
     public void test_namefilter_1() throws Exception {
@@ -56,23 +54,23 @@ public class NameFilterTest_char extends TestCase {
         serializer.getNameFilters().add(filter);
 
         Map<String, Object> map = new HashMap<String, Object>();
-        map.put("id", '0');
+        map.put("id", 0);
         serializer.write(map);
 
         String text = out.toString();
-        Assert.assertEquals("{\"ID\":\"0\"}", text);
+        Assert.assertEquals("{\"ID\":0}", text);
     }
 
     public static class Bean {
 
-        private char   id = '0';
+        private double  id;
         private String name;
 
-        public char getId() {
+        public double getId() {
             return id;
         }
 
-        public void setId(char id) {
+        public void setId(double id) {
             this.id = id;
         }
