+++ src/test/java/com/alibaba/json/bvt/serializer/filters/NameFilterTest_float.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt.serializer;
+package com.alibaba.json.bvt.serializer.filters;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -10,14 +10,13 @@ import com.alibaba.fastjson.serializer.JSONSerializer;
 import com.alibaba.fastjson.serializer.NameFilter;
 import com.alibaba.fastjson.serializer.SerializeWriter;
 
-public class NameFilterTest_long extends TestCase {
+public class NameFilterTest_float extends TestCase {
 
     public void test_namefilter() throws Exception {
         NameFilter filter = new NameFilter() {
 
             public String process(Object source, String name, Object value) {
                 if (name.equals("id")) {
-                    Assert.assertTrue(value instanceof Long);
                     return "ID";
                 }
 
@@ -64,14 +63,14 @@ public class NameFilterTest_long extends TestCase {
 
     public static class Bean {
 
-        private long   id;
+        private float  id;
         private String name;
 
-        public long getId() {
+        public float getId() {
             return id;
         }
 
-        public void setId(long id) {
+        public void setId(float id) {
             this.id = id;
         }
