+++ src/test/java/com/alibaba/json/bvt/serializer/filters/NameFilterTest_short.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt.serializer;
+package com.alibaba.json.bvt.serializer.filters;
 
 import java.util.HashMap;
 import java.util.Map;
